+++ app/(auth)/signin/page.tsx
@@ -1,6 +1,14 @@
 import { redirect } from "next/navigation";
 import { createClient } from "@/utils/supabase/server";
 import SignInComponent from "@/components/auth/signin";
+import { constructMetadata } from "@/lib/utils";
+import { Metadata } from "next/types";
+
+export const metadata: Metadata = constructMetadata({
+  title: "Sign In",
+  description: "Sign in to your account",
+  canonical: "/signin",
+});
 
 export default async function SignIn() {
   const supabase = createClient();
